0.6.0 -- use one connection per worker
See discussion starting at
https://gitter.im/contribsys/faktory?at=5a007ad332e080696e6f6fe4.
In essence, the issue is that if you have many workers and few jobs, the
ACK of a job will be delayed by blocking FETCHes by other workers in the
old model. In the new model, this is not an issue. It does, however,
make the testing a bit more complex, but that seems worth it.

This change also makes it so that the connector and the url are cached
inside `Client` , which makes for a much nicer reconnect interface.