Remove TLS support
Now that https://github.com/contribsys/faktory/pull/76 has been merged,
there is no need to have explicit TLS support in the client. This makes
the interface *much* nicer, as we can just assume TCP. We still provide
a way of using other transports (for testing, or if we eventually get
support for pipes or unix sockets), but that's now separate.