Rework consumer type to use function-like-object via a trait instead of closures.
It seems to actually work quite smoothly, and you can still pass closures
into it to create job runners.  So, this is strictly more powerful, I think.

If you do everything *exactly right* then it's possible to work sanely
with closures, but it's really hard to figure it out in general.